can `npm test` web socket
