+++ src/socket/web/test/client.ts
@@ -1,120 +1,30 @@
-import ip = require('ip');
-import test = require('tape');
+import tcp = require('net');
+import { spawn } from 'child_process';
 
-import { MuSocketState } from '../../../socket';
-import { MuWebSocket } from '../client';
-
-function noop () { }
-
-function id () {
-    return Math.random().toString(36).substr(2);
+function getFreePort (cb:(port:number) => void) {
+    const server = tcp.createServer();
+    server.on('error', (e) => console.error(e));
+    server.unref();
+    server.listen(() => {
+        const port = server.address().port;
+        server.close(() => cb(port));
+    });
 }
 
-const url = `ws://${ip.address()}:${process.env.PORT}`;
+getFreePort((port) => {
+    const cwd = __dirname;
 
-if (typeof process === 'object' && 'exit' in process) {
-    test.onFinish(() => process.exit(0));
-}
+    const server = spawn(
+        'ts-node',
+        [ '_/start-server.ts', `${port}` ],
+        { cwd },
+    );
+    server.stdout.once('data', (data) => console.log(`${data}`));
 
-test('socket initial state', (t) => {
-    const socket = new MuWebSocket({
-        sessionId: id(),
-        url,
+    const test = spawn(`browserify ${cwd}/_/client.ts -p [ tsify ] -t [ envify --PORT ${port} ] | tape-run`, [], {
+        cwd,
+        shell: true,
+        stdio: 'inherit',
     });
-    t.equals(socket.state, MuSocketState.INIT, 'should be INIT');
-    t.end();
-});
-
-test('socket.open() - when INIT', (t) => {
-    t.plan(1);
-
-    let callsToReady = 0;
-
-    const socket = new MuWebSocket({
-        sessionId: id(),
-        url,
-    });
-
-    socket.open({
-        ready: () => {
-            ++callsToReady;
-
-            if (callsToReady > 1) {
-                t.fail('should not call ready handler more than once');
-            }
-            t.equals(socket.state, MuSocketState.OPEN, 'should change socket state to OPEN');
-        },
-        message: noop,
-        close: noop,
-    });
-});
-
-test('socket.open() - when OPEN', (t) => {
-    t.plan(1);
-
-    const socket = new MuWebSocket({
-        sessionId: id(),
-        url,
-    });
-
-    socket.open({
-        ready: () => {
-            t.throws(
-                () => socket.open({
-                    ready: noop,
-                    message: noop,
-                    close: noop,
-                }),
-            );
-        },
-        message: noop,
-        close: noop,
-    });
-});
-
-test('socket.open() - when CLOSED', (t) => {
-    t.plan(1);
-
-    const socket = new MuWebSocket({
-        sessionId: id(),
-        url,
-    });
-
-    socket.open({
-        ready: () => socket.close(),
-        message: noop,
-        close: () => t.throws(
-            () => socket.open({
-                ready: noop,
-                message: noop,
-                close: noop,
-            }),
-        ),
-    });
-});
-
-test('socket.close() - when OPEN', (t) => {
-    t.plan(1);
-
-    const socket = new MuWebSocket({
-        sessionId: id(),
-        url,
-    });
-
-    socket.open({
-        ready: () => socket.close(),
-        message: noop,
-        close: () => t.equals(socket.state, MuSocketState.CLOSED, 'should change socket state to CLOSED'),
-    });
-});
-
-test('socket.close() - when INIT', (t) => {
-    const socket = new MuWebSocket({
-        sessionId: id(),
-        url,
-    });
-
-    socket.close();
-    t.equals(socket.state, MuSocketState.CLOSED, 'should change socket state to CLOSED');
-    t.end();
+    test.on('exit', process.exit);
 });
