+++ src/socket/web/test/server.ts
@@ -1,8 +1,8 @@
-import http = require('http');
-
 import test = require('tape');
 import ip = require('ip');
 import WebSocket = require('uws');
+
+import http = require('http');
 
 import { MuSocketServerState } from '../../../socket';
 import { MuWebSocketServer } from '../server';
@@ -15,7 +15,7 @@ test.onFinish(() => process.exit(0));
 
 test('server initial state', (t) => {
     const socketServer = new MuWebSocketServer({ server });
-    t.equals(socketServer.state, MuSocketServerState.INIT, 'should be INIT');
+    t.equal(socketServer.state, MuSocketServerState.INIT, 'should be INIT');
     t.end();
 });
 
@@ -26,7 +26,7 @@ test('socketServer.start() - when INIT', (t) => {
     socketServer.start({
         ready: () => {
             t.pass('should invoke ready handler');
-            t.equals(socketServer.state, MuSocketServerState.RUNNING, 'should change server state to RUNNING');
+            t.equal(socketServer.state, MuSocketServerState.RUNNING, 'should change server state to RUNNING');
         },
         connection: noop,
         close: noop,
@@ -138,21 +138,20 @@ test('socketServer.close() - when INIT', (t) => {
     const socketServer = new MuWebSocketServer({ server });
     socketServer.close();
 
-    t.equals(socketServer.state, MuSocketServerState.SHUTDOWN, 'should change server state to SHUTDOWN');
+    t.equal(socketServer.state, MuSocketServerState.SHUTDOWN, 'should change server state to SHUTDOWN');
     t.end();
 });
 
 test('socketServer.close() - when RUNNING', (t) => {
+    t.plan(1);
+
     const socketServer = new MuWebSocketServer({ server });
     socketServer.start({
         ready: () => {
             socketServer.close();
-            t.equals(socketServer.state, MuSocketServerState.SHUTDOWN, 'should change server state to SHUTDOWN');
+            t.equal(socketServer.state, MuSocketServerState.SHUTDOWN, 'should change server state to SHUTDOWN');
         },
         connection: noop,
-        close: (error) => {
-            t.equals(error, undefined, 'should invoke close handler without error message');
-            t.end();
-        },
+        close: noop,
     });
 });
